Flatten providers subpackage
