+++ docs/apidocs/providers.rst
@@ -1,10 +1,5 @@
 .. _qiskit-providers:
 
-.. automodule:: qiskit.providers.v2
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
-
 .. automodule:: qiskit.providers
    :no-members:
    :no-inherited-members:
