+++ qiskit/compiler/assemble.py
@@ -26,7 +26,7 @@ from qiskit.qobj import QobjHeader, Qobj
 from qiskit.qobj.utils import MeasLevel, MeasReturnType
 from qiskit.validation.jsonschema import SchemaValidationError
 from qiskit.providers import BaseBackend
-from qiskit.providers.v2 import Backend
+from qiskit.providers.backend import Backend
 from qiskit.pulse.channels import PulseChannel
 from qiskit.pulse import Schedule
